feat: unify agent interface
