+++ src/agents/core/__init__.py
@@ -1,0 +1,3 @@
+"""Core agent interfaces and utilities."""
+
+from .agent_interface import *  # noqa: F403
