+++ src/core/agents/specialist_agent_registry.py
@@ -17,6 +17,8 @@ from enum import Enum
 from pathlib import Path
 from typing import Any
 
+from agents.core.agent_interface import AgentCapability
+
 logger = logging.getLogger(__name__)
 
 
@@ -30,29 +32,6 @@ class AgentRole(Enum):
     COMMUNICATION = "communication"  # Herald, Emissary, Scribe
     MANAGEMENT = "management"  # Steward, Curator, Logger
     ANALYSIS = "analysis"  # Profiler
-
-
-class AgentCapability(Enum):
-    """Specific capabilities agents can provide."""
-
-    DECISION_MAKING = "decision_making"
-    STRATEGIC_PLANNING = "strategic_planning"
-    THREAT_DETECTION = "threat_detection"
-    DEFENSE_COORDINATION = "defense_coordination"
-    KNOWLEDGE_SYNTHESIS = "knowledge_synthesis"
-    PATH_OPTIMIZATION = "path_optimization"
-    RESOURCE_ALLOCATION = "resource_allocation"
-    TASK_ORCHESTRATION = "task_orchestration"
-    COMMUNICATION_ROUTING = "communication_routing"
-    MESSAGE_ENCODING = "message_encoding"
-    DATA_PERSISTENCE = "data_persistence"
-    EVENT_LOGGING = "event_logging"
-    PERFORMANCE_ANALYSIS = "performance_analysis"
-    CONTENT_CURATION = "content_curation"
-    ASSET_TRANSFORMATION = "asset_transformation"
-    SYSTEM_CONSTRUCTION = "system_construction"
-    HISTORICAL_ANALYSIS = "historical_analysis"
-    DIPLOMATIC_NEGOTIATION = "diplomatic_negotiation"
 
 
 class AgentStatus(Enum):
@@ -128,7 +107,9 @@ class AgentInstance:
 
     # Network state
     endpoint: str | None = None
-    transport_preferences: list[str] = field(default_factory=lambda: ["betanet", "bitchat"])
+    transport_preferences: list[str] = field(
+        default_factory=lambda: ["betanet", "bitchat"]
+    )
     latency_ms: float = 0.0
 
 
@@ -151,7 +132,10 @@ class SpecialistAgentRegistry:
         self._initialize_core_agent_specs()
         self._build_capability_indexes()
 
-        logger.info(f"SpecialistAgentRegistry initialized with {len(self.specifications)} agent types")
+        logger.info(
+            "SpecialistAgentRegistry initialized with "
+            f"{len(self.specifications)} agent types"
+        )
 
     def _initialize_core_agent_specs(self):
         """Initialize specifications for all 18 core agents."""
@@ -426,7 +410,9 @@ class SpecialistAgentRegistry:
                     self.capability_index[capability] = []
                 self.capability_index[capability].append(agent_type)
 
-    def register_instance(self, agent_type: str, device_id: str, **kwargs) -> AgentInstance:
+    def register_instance(
+        self, agent_type: str, device_id: str, **kwargs
+    ) -> AgentInstance:
         """Register a new agent instance."""
         if agent_type not in self.specifications:
             raise ValueError(f"Unknown agent type: {agent_type}")
@@ -444,7 +430,9 @@ class SpecialistAgentRegistry:
         )
 
         self.instances[instance_id] = instance
-        logger.info(f"Registered {agent_type} instance {instance_id} on device {device_id}")
+        logger.info(
+            f"Registered {agent_type} instance {instance_id} on device {device_id}"
+        )
 
         return instance
 
@@ -504,7 +492,9 @@ class SpecialistAgentRegistry:
             candidates = self.get_available_instances()
 
         if not candidates:
-            logger.warning(f"No available agents for task: {task.get('task_id', 'unknown')}")
+            logger.warning(
+                f"No available agents for task: {task.get('task_id', 'unknown')}"
+            )
             return None
 
         # Score candidates based on suitability
@@ -518,7 +508,9 @@ class SpecialistAgentRegistry:
             score = spec.collaboration_score * spec.priority_multiplier
 
             # Adjust for current load
-            load_factor = 1.0 - (len(candidate.current_tasks) / spec.max_concurrent_tasks)
+            load_factor = 1.0 - (
+                len(candidate.current_tasks) / spec.max_concurrent_tasks
+            )
             score *= load_factor
 
             # Adjust for performance history
@@ -526,7 +518,9 @@ class SpecialistAgentRegistry:
             score *= success_rate
 
             # Adjust for capability match
-            if required_capability and required_capability in [c.value for c in candidate.capabilities_available]:
+            if required_capability and required_capability in [
+                c.value for c in candidate.capabilities_available
+            ]:
                 score *= 1.2
 
             if score > best_score:
@@ -537,14 +531,19 @@ class SpecialistAgentRegistry:
             # Assign task to agent
             task_id = task.get("task_id", str(uuid.uuid4()))
             best_candidate.current_tasks.append(task_id)
-            # Only mark as busy if at max capacity, otherwise keep available for more tasks
+            # Mark busy only if at capacity; otherwise keep available for more tasks
             spec = self.specifications[best_candidate.agent_type]
             if len(best_candidate.current_tasks) >= spec.max_concurrent_tasks:
                 best_candidate.status = AgentStatus.BUSY
             else:
                 best_candidate.status = AgentStatus.AVAILABLE
 
-            logger.info(f"Routed task {task_id} to {best_candidate.agent_type} instance {best_candidate.instance_id}")
+            logger.info(
+                "Routed task %s to %s instance %s",
+                task_id,
+                best_candidate.agent_type,
+                best_candidate.instance_id,
+            )
 
         return best_candidate
 
@@ -562,17 +561,22 @@ class SpecialistAgentRegistry:
         # Count instances by status
         for instance in self.instances.values():
             status_key = instance.status.value
-            status["instances_by_status"][status_key] = status["instances_by_status"].get(status_key, 0) + 1
+            status["instances_by_status"][status_key] = (
+                status["instances_by_status"].get(status_key, 0) + 1
+            )
 
             type_key = instance.agent_type
-            status["instances_by_type"][type_key] = status["instances_by_type"].get(type_key, 0) + 1
+            status["instances_by_type"][type_key] = (
+                status["instances_by_type"].get(type_key, 0) + 1
+            )
 
         # Count capability coverage
         for capability, agent_types in self.capability_index.items():
             available_instances = sum(
                 1
                 for instance in self.instances.values()
-                if instance.agent_type in agent_types and instance.status == AgentStatus.AVAILABLE
+                if instance.agent_type in agent_types
+                and instance.status == AgentStatus.AVAILABLE
             )
             status["capability_coverage"][capability.value] = {
                 "agent_types": len(agent_types),
@@ -584,7 +588,8 @@ class SpecialistAgentRegistry:
             available_instances = sum(
                 1
                 for instance in self.instances.values()
-                if instance.agent_type in agent_types and instance.status == AgentStatus.AVAILABLE
+                if instance.agent_type in agent_types
+                and instance.status == AgentStatus.AVAILABLE
             )
             status["role_coverage"][role.value] = {
                 "agent_types": len(agent_types),
@@ -601,7 +606,9 @@ class SpecialistAgentRegistry:
             export_data[agent_type] = asdict(spec)
             # Convert enums to strings for JSON serialization
             export_data[agent_type]["primary_role"] = spec.primary_role.value
-            export_data[agent_type]["secondary_roles"] = [r.value for r in spec.secondary_roles]
+            export_data[agent_type]["secondary_roles"] = [
+                r.value for r in spec.secondary_roles
+            ]
 
             # Convert capability configs
             for cap_config in export_data[agent_type]["capabilities"]:
