+++ src/production/rag/rag_system/core/agent_interface.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for the unified AgentInterface."""
+
+from agents.core.agent_interface import *  # noqa: F403
